+++ dbms/src/Flash/tests/gtest_spill_sort.cpp
@@ -60,7 +60,7 @@ try
     enablePipeline(false);
     /// disable spill
     context.context->setSetting("max_bytes_before_external_sort", Field(static_cast<UInt64>(0)));
-    auto ref_columns = executeStreams(request, original_max_streams);
+    auto ref_columns = executeStreams(request, 1);
     /// enable spill
     context.context->setSetting("max_bytes_before_external_sort", Field(static_cast<UInt64>(total_data_size / 10)));
     // don't use `executeAndAssertColumnsEqual` since it takes too long to run
@@ -72,9 +72,9 @@ try
 
     // The implementation of topN in the pipeline model is LocalSort, and the result of using multiple threads is unstable. Therefore, a single thread is used here instead.
     enablePipeline(true);
-    ASSERT_COLUMNS_EQ_UR(ref_columns, executeStreams(request, 1));
+    ASSERT_COLUMNS_EQ_R(ref_columns, executeStreams(request, 1));
     context.context->setSetting("max_cached_data_bytes_in_spiller", Field(static_cast<UInt64>(total_data_size / 100)));
-    ASSERT_COLUMNS_EQ_UR(ref_columns, executeStreams(request, 1));
+    ASSERT_COLUMNS_EQ_R(ref_columns, executeStreams(request, 1));
 }
 CATCH
 
@@ -110,7 +110,7 @@ try
         enablePipeline(false);
         /// disable spill
         context.context->setSetting("max_bytes_before_external_sort", Field(static_cast<UInt64>(0)));
-        auto ref_columns = executeStreams(request, original_max_streams);
+        auto ref_columns = executeStreams(request, 1);
         /// enable spill
         context.context->setSetting("max_bytes_before_external_sort", Field(static_cast<UInt64>(total_data_size / 10)));
         // don't use `executeAndAssertColumnsEqual` since it takes too long to run
@@ -122,9 +122,9 @@ try
 
         // The implementation of topN in the pipeline model is LocalSort, and the result of using multiple threads is unstable. Therefore, a single thread is used here instead.
         enablePipeline(true);
-        ASSERT_COLUMNS_EQ_UR(ref_columns, executeStreams(request, 1));
+        ASSERT_COLUMNS_EQ_R(ref_columns, executeStreams(request, 1));
         context.context->setSetting("max_cached_data_bytes_in_spiller", Field(static_cast<UInt64>(total_data_size / 100)));
-        ASSERT_COLUMNS_EQ_UR(ref_columns, executeStreams(request, 1));
+        ASSERT_COLUMNS_EQ_R(ref_columns, executeStreams(request, 1));
     }
 }
 CATCH
